requirements.txt
streamlit==1.22.0
streamlit-image-select==0.6.0
pillow==9.5.0